restructure e2e tests
